+++ src/create.spec.ts
@@ -6,7 +6,7 @@ describe("js-rln", () => {
   it("should verify a proof", async function () {
     const rlnInstance = await createRLN();
 
-    const credential = rlnInstance.generateIdentityCredentials();
+    const credential = rlnInstance.zerokit.generateIdentityCredentials();
 
     //peer's index in the Merkle Tree
     const index = 5;
@@ -15,11 +15,11 @@ describe("js-rln", () => {
     for (let i = 0; i < 10; i++) {
       if (i == index) {
         // insert the current peer's pk
-        rlnInstance.insertMember(credential.IDCommitment);
+        rlnInstance.zerokit.insertMember(credential.IDCommitment);
       } else {
         // create a new key pair
-        rlnInstance.insertMember(
-          rlnInstance.generateIdentityCredentials().IDCommitment
+        rlnInstance.zerokit.insertMember(
+          rlnInstance.zerokit.generateIdentityCredentials().IDCommitment
         );
       }
     }
@@ -33,7 +33,7 @@ describe("js-rln", () => {
     const epoch = new Date();
 
     // generating proof
-    const proof = await rlnInstance.generateRLNProof(
+    const proof = await rlnInstance.zerokit.generateRLNProof(
       uint8Msg,
       index,
       epoch,
@@ -42,7 +42,7 @@ describe("js-rln", () => {
 
     try {
       // verify the proof
-      const verifResult = rlnInstance.verifyRLNProof(proof, uint8Msg);
+      const verifResult = rlnInstance.zerokit.verifyRLNProof(proof, uint8Msg);
       expect(verifResult).to.be.true;
     } catch (err) {
       assert.fail(0, 1, "should not have failed proof verification");
@@ -52,7 +52,7 @@ describe("js-rln", () => {
       // Modifying the signal so it's invalid
       uint8Msg[4] = 4;
       // verify the proof
-      const verifResult = rlnInstance.verifyRLNProof(proof, uint8Msg);
+      const verifResult = rlnInstance.zerokit.verifyRLNProof(proof, uint8Msg);
       expect(verifResult).to.be.false;
     } catch (err) {
       console.log(err);
@@ -61,7 +61,8 @@ describe("js-rln", () => {
   it("should verify a proof with a seeded membership key generation", async function () {
     const rlnInstance = await createRLN();
     const seed = "This is a test seed";
-    const credential = rlnInstance.generateSeededIdentityCredential(seed);
+    const credential =
+      rlnInstance.zerokit.generateSeededIdentityCredential(seed);
 
     //peer's index in the Merkle Tree
     const index = 5;
@@ -70,11 +71,11 @@ describe("js-rln", () => {
     for (let i = 0; i < 10; i++) {
       if (i == index) {
         // insert the current peer's pk
-        rlnInstance.insertMember(credential.IDCommitment);
+        rlnInstance.zerokit.insertMember(credential.IDCommitment);
       } else {
         // create a new key pair
-        rlnInstance.insertMember(
-          rlnInstance.generateIdentityCredentials().IDCommitment
+        rlnInstance.zerokit.insertMember(
+          rlnInstance.zerokit.generateIdentityCredentials().IDCommitment
         );
       }
     }
@@ -88,7 +89,7 @@ describe("js-rln", () => {
     const epoch = new Date();
 
     // generating proof
-    const proof = await rlnInstance.generateRLNProof(
+    const proof = await rlnInstance.zerokit.generateRLNProof(
       uint8Msg,
       index,
       epoch,
@@ -97,7 +98,7 @@ describe("js-rln", () => {
 
     try {
       // verify the proof
-      const verifResult = rlnInstance.verifyRLNProof(proof, uint8Msg);
+      const verifResult = rlnInstance.zerokit.verifyRLNProof(proof, uint8Msg);
       expect(verifResult).to.be.true;
     } catch (err) {
       assert.fail(0, 1, "should not have failed proof verification");
@@ -107,7 +108,7 @@ describe("js-rln", () => {
       // Modifying the signal so it's invalid
       uint8Msg[4] = 4;
       // verify the proof
-      const verifResult = rlnInstance.verifyRLNProof(proof, uint8Msg);
+      const verifResult = rlnInstance.zerokit.verifyRLNProof(proof, uint8Msg);
       expect(verifResult).to.be.false;
     } catch (err) {
       console.log(err);
@@ -117,8 +118,8 @@ describe("js-rln", () => {
   it("should generate the same membership key if the same seed is provided", async function () {
     const rlnInstance = await createRLN();
     const seed = "This is a test seed";
-    const memKeys1 = rlnInstance.generateSeededIdentityCredential(seed);
-    const memKeys2 = rlnInstance.generateSeededIdentityCredential(seed);
+    const memKeys1 = rlnInstance.zerokit.generateSeededIdentityCredential(seed);
+    const memKeys2 = rlnInstance.zerokit.generateSeededIdentityCredential(seed);
 
     memKeys1.IDCommitment.forEach((element, index) => {
       expect(element).to.equal(memKeys2.IDCommitment[index]);
