+++ src/rln.ts
@@ -2,7 +2,6 @@ import { createDecoder, createEncoder } from "@waku/core";
 import type {
   ContentTopic,
   IDecodedMessage,
-  IRateLimitProof,
   EncoderOptions as WakuEncoderOptions,
 } from "@waku/interfaces";
 import init from "@waku/zerokit-rln-wasm";
@@ -23,17 +22,11 @@ import type {
   EncryptedCredentials,
 } from "./keystore/index.js";
 import { KeystoreEntity, Password } from "./keystore/types.js";
-import { Proof, proofToBytes } from "./proof.js";
 import verificationKey from "./resources/verification_key.js";
 import * as wc from "./resources/witness_calculator.js";
 import { WitnessCalculator } from "./resources/witness_calculator.js";
-import {
-  concatenate,
-  dateToEpoch,
-  epochIntToBytes,
-  extractMetaMaskSigner,
-  writeUIntLE,
-} from "./utils/index.js";
+import { extractMetaMaskSigner } from "./utils/index.js";
+import { Zerokit } from "./zerokit.js";
 
 async function loadWitnessCalculator(): Promise<WitnessCalculator> {
   const url = new URL("./resources/rln.wasm", import.meta.url);
@@ -63,8 +56,9 @@ export async function create(): Promise<RLNInstance> {
 
   const DEPTH = 20;
   const zkRLN = zerokitRLN.newRLN(DEPTH, zkey, vkey);
+  const zerokit = new Zerokit(zkRLN, witnessCalculator);
 
-  return new RLNInstance(zkRLN, witnessCalculator);
+  return new RLNInstance(zerokit);
 }
 
 type StartRLNOptions = {
@@ -101,10 +95,7 @@ export class RLNInstance {
   private keystore = Keystore.create();
   private _credentials: undefined | DecryptedCredentials;
 
-  constructor(
-    private zkRLN: number,
-    private witnessCalculator: WitnessCalculator
-  ) {}
+  constructor(public zerokit: Zerokit) {}
 
   public get contract(): undefined | RLNContract {
     return this._contract;
@@ -212,7 +203,9 @@ export class RLNInstance {
     let identity = "identity" in options && options.identity;
 
     if ("signature" in options) {
-      identity = await this.generateSeededIdentityCredential(options.signature);
+      identity = await this.zerokit.generateSeededIdentityCredential(
+        options.signature
+      );
     }
 
     if (!identity) {
@@ -288,167 +281,5 @@ export class RLNInstance {
       rlnInstance: this,
       decoder: createDecoder(contentTopic),
     });
-  }
-
-  generateIdentityCredentials(): IdentityCredential {
-    const memKeys = zerokitRLN.generateExtendedMembershipKey(this.zkRLN); // TODO: rename this function in zerokit rln-wasm
-    return IdentityCredential.fromBytes(memKeys);
-  }
-
-  generateSeededIdentityCredential(seed: string): IdentityCredential {
-    const stringEncoder = new TextEncoder();
-    const seedBytes = stringEncoder.encode(seed);
-    // TODO: rename this function in zerokit rln-wasm
-    const memKeys = zerokitRLN.generateSeededExtendedMembershipKey(
-      this.zkRLN,
-      seedBytes
-    );
-    return IdentityCredential.fromBytes(memKeys);
-  }
-
-  insertMember(idCommitment: Uint8Array): void {
-    zerokitRLN.insertMember(this.zkRLN, idCommitment);
-  }
-
-  insertMembers(index: number, ...idCommitments: Array<Uint8Array>): void {
-    // serializes a seq of IDCommitments to a byte seq
-    // the order of serialization is |id_commitment_len<8>|id_commitment<var>|
-    const idCommitmentLen = writeUIntLE(
-      new Uint8Array(8),
-      idCommitments.length,
-      0,
-      8
-    );
-    const idCommitmentBytes = concatenate(idCommitmentLen, ...idCommitments);
-    zerokitRLN.setLeavesFrom(this.zkRLN, index, idCommitmentBytes);
-  }
-
-  deleteMember(index: number): void {
-    zerokitRLN.deleteLeaf(this.zkRLN, index);
-  }
-
-  getMerkleRoot(): Uint8Array {
-    return zerokitRLN.getRoot(this.zkRLN);
-  }
-
-  serializeMessage(
-    uint8Msg: Uint8Array,
-    memIndex: number,
-    epoch: Uint8Array,
-    idKey: Uint8Array
-  ): Uint8Array {
-    // calculate message length
-    const msgLen = writeUIntLE(new Uint8Array(8), uint8Msg.length, 0, 8);
-
-    // Converting index to LE bytes
-    const memIndexBytes = writeUIntLE(new Uint8Array(8), memIndex, 0, 8);
-
-    // [ id_key<32> | id_index<8> | epoch<32> | signal_len<8> | signal<var> ]
-    return concatenate(idKey, memIndexBytes, epoch, msgLen, uint8Msg);
-  }
-
-  async generateRLNProof(
-    msg: Uint8Array,
-    index: number,
-    epoch: Uint8Array | Date | undefined,
-    idSecretHash: Uint8Array
-  ): Promise<IRateLimitProof> {
-    if (epoch == undefined) {
-      epoch = epochIntToBytes(dateToEpoch(new Date()));
-    } else if (epoch instanceof Date) {
-      epoch = epochIntToBytes(dateToEpoch(epoch));
-    }
-
-    if (epoch.length != 32) throw "invalid epoch";
-    if (idSecretHash.length != 32) throw "invalid id secret hash";
-    if (index < 0) throw "index must be >= 0";
-
-    const serialized_msg = this.serializeMessage(
-      msg,
-      index,
-      epoch,
-      idSecretHash
-    );
-    const rlnWitness = zerokitRLN.getSerializedRLNWitness(
-      this.zkRLN,
-      serialized_msg
-    );
-    const inputs = zerokitRLN.RLNWitnessToJson(this.zkRLN, rlnWitness);
-    const calculatedWitness = await this.witnessCalculator.calculateWitness(
-      inputs,
-      false
-    ); // no sanity check being used in zerokit
-
-    const proofBytes = zerokitRLN.generate_rln_proof_with_witness(
-      this.zkRLN,
-      calculatedWitness,
-      rlnWitness
-    );
-
-    return new Proof(proofBytes);
-  }
-
-  verifyRLNProof(
-    proof: IRateLimitProof | Uint8Array,
-    msg: Uint8Array
-  ): boolean {
-    let pBytes: Uint8Array;
-    if (proof instanceof Uint8Array) {
-      pBytes = proof;
-    } else {
-      pBytes = proofToBytes(proof);
-    }
-
-    // calculate message length
-    const msgLen = writeUIntLE(new Uint8Array(8), msg.length, 0, 8);
-
-    return zerokitRLN.verifyRLNProof(
-      this.zkRLN,
-      concatenate(pBytes, msgLen, msg)
-    );
-  }
-
-  verifyWithRoots(
-    proof: IRateLimitProof | Uint8Array,
-    msg: Uint8Array,
-    ...roots: Array<Uint8Array>
-  ): boolean {
-    let pBytes: Uint8Array;
-    if (proof instanceof Uint8Array) {
-      pBytes = proof;
-    } else {
-      pBytes = proofToBytes(proof);
-    }
-    // calculate message length
-    const msgLen = writeUIntLE(new Uint8Array(8), msg.length, 0, 8);
-
-    const rootsBytes = concatenate(...roots);
-
-    return zerokitRLN.verifyWithRoots(
-      this.zkRLN,
-      concatenate(pBytes, msgLen, msg),
-      rootsBytes
-    );
-  }
-
-  verifyWithNoRoot(
-    proof: IRateLimitProof | Uint8Array,
-    msg: Uint8Array
-  ): boolean {
-    let pBytes: Uint8Array;
-    if (proof instanceof Uint8Array) {
-      pBytes = proof;
-    } else {
-      pBytes = proofToBytes(proof);
-    }
-
-    // calculate message length
-    const msgLen = writeUIntLE(new Uint8Array(8), msg.length, 0, 8);
-
-    return zerokitRLN.verifyWithRoots(
-      this.zkRLN,
-      concatenate(pBytes, msgLen, msg),
-      new Uint8Array()
-    );
   }
 }
